+++ src/pages/SettingsView.spec.ts
@@ -32,10 +32,10 @@ describe(SettingsView.name, () => {
   })
 
   describe('when page loads', () => {
-    it('shows general settings', async () => {
+    it('hides general settings', async () => {
       await renderComponent()
       const field = screen.getByLabelText('Device name')
-      expect(field).toBeVisible()
+      expect(field).not.toBeVisible()
     })
 
     it('hides camera settings', async () => {
@@ -46,6 +46,92 @@ describe(SettingsView.name, () => {
 
     it('hides trigger settings', async () => {
       await renderComponent()
+      const field = screen.getByText('Trigger sensitivity')
+      expect(field).not.toBeVisible()
+    })
+  })
+
+  describe('when general settings heading is clicked', () => {
+    const user = userEvent.setup()
+
+    it('shows general settings', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
+      const field = screen.getByLabelText('Device name')
+      expect(field).toBeVisible()
+    })
+
+    it('shows site name setting', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
+      const field = screen.getByText('Site name (optional)')
+      expect(field).toBeVisible()
+    })
+
+    it('shows device name setting', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
+      const field = screen.getByText('Device name')
+      expect(field).toBeVisible()
+    })
+
+    it('shows device id setting', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
+      const field = screen.getByText('Device ID')
+      expect(field).toBeVisible()
+    })
+
+    it('shows time zone setting', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
+      const field = screen.getByText('Time zone')
+      expect(field).toBeVisible()
+    })
+
+    it('shows date setting', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
+      const field = screen.getByText('Date')
+      expect(field).toBeVisible()
+    })
+
+    it('shows time setting', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
+      const field = screen.getByText('Time')
+      expect(field).toBeVisible()
+    })
+
+    it('shows filename preview heading', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
+      const heading = screen.queryByRole('heading', {
+        name: 'Filename preview',
+      })
+      expect(heading).toBeVisible()
+    })
+
+    it('hides camera settings', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
+      const field = screen.getByLabelText('Pictures')
+      expect(field).not.toBeVisible()
+    })
+
+    it('hides trigger settings', async () => {
+      await renderComponent()
+      const button = screen.getByRole('button', { name: /General settings/ })
+      await user.click(button)
       const field = screen.getByText('Trigger sensitivity')
       expect(field).not.toBeVisible()
     })